backend/merge_videos/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
moviepy